requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
flet>=0.21.0
pyperclip>=1.8.0
keyboard>=0.13.5
requests>=2.28.0
